\name{graphicsqc-package}
\alias{graphicsqc-package}
\alias{graphicsqc}
\docType{package}
\title{
Quality Control for Graphics
}
\description{
Generates graphics files, compares them with "model" files,
and reports the results.
}
\details{
\tabular{ll}{
Package: \tab graphicsqc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-07-24\cr
License: \tab GPL-2\cr
}
Use the convenience 'graphicsQC' function..
}
\author{
Stephen Gardiner and Paul Murrell

Maintainer: Paul Murrell <paul@stat.auckland.ac.nz>
}
\references{
Murrell, P. & Hornik, K. 2003 Quality Assurance for Graphics in R
\url{http://www.ci.tuwien.ac.at/Conferences/DSC-2003/Proceedings/MurrellHornik.pdf}
}
\keyword{package}
\examples{
#graphicsQC(...)
}
