\name{graphicsqc-package}
\alias{graphicsqc-package}
\alias{graphicsqc}
\docType{package}
\title{
Quality Control for Graphics
}
\description{
Generates graphics files, compares them with "model" files,
and reports the results.
}
\details{
\tabular{ll}{
Package: \tab graphicsqc\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2008-07-24\cr
License: \tab GPL-2\cr
}
Use the convenience 'graphicsQC' function..
}
\author{
Stephen Gardiner

Maintainer: Paul Murrell <paul@stat.auckland.ac.nz>
}
\references{
Free Software Foundation, Inc. 2008 \emph{Diffutils}. <\url{http://www.gnu.org/software/diffutils/diffutils.html}> 

ImageMagick Studio LLC. 2008 \emph{ImageMagick}. <\url{http://www.imagemagick.org/}>.

Murrell, P. & Hornik, K. 2003 \emph{Quality Assurance for Graphics in \R} <\url{http://www.ci.tuwien.ac.at/Conferences/DSC-2003/Proceedings/MurrellHornik.pdf}>.
}
\seealso{
\code{plotExpr}, \code{plotFun}
}
\examples{
#graphicsQC(...)
}
\keyword{utilities}
